feat: stargazers and forks query added
